# apiDados/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api Hypatiamat",
        "description": "Aqui encontra-se a documentação de toda a api de dados do Hypatiamat, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3050",
    "basePath": "/",
...
    "paths": {
...
        "/aplicacoes/professores/{idProf}/turmas": {
            "get": {
              "tags": [
                "Professores"
              ],
              "summary": "Dado um identificador de um professor, devolve todas as suas turmas.",
              "description": "Devolve todos as turmas de um professor.",
              "operationId": "getTurmasProfessorByCodigo",
              "parameters": [
                {
                  "name": "idProf",
                  "in": "path",
                  "description": "Id do professor a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                      "description": "Lista de Turmas",
                      "schema": {
                        "type":"array",
                        "items":{
                          "$ref": "#/definitions/Turma"
                        }
                      }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
...
        "/aplicacoes/alunos/{idAluno}/turma":{
          "put":{
            "tags": [
              "Alunos"
            ],
            "summary": "Dado um identificador de um aluno, edite a password dele.",
            "description": "Editar a password de um aluno.",
            "operationId": "putPasswordAlunoById",
            "parameters": [
              {
                "name": "idAluno",
                "in": "path",
                "description": "Identificador de um aluno.",
                "required": true,
                "type": "integer"
              },
              {
                "name": "turma",
                "in": "formData",
                "description": "Nova turma do aluno.",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "A turma do aluno foi alterada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
              "security": [
                {
                  "api_key": []
                }
              ]
          }
        },
...
      "/aplicacoes/turmas/":{
        "get": {
          "tags": [
            "Turmas"
          ],
          "summary": "Devolve todas as turmas do sistema.",
          "description": "Retorna todas as turmas registadas.",
          "operationId": "getTurmas",
          "responses": {
            "200": {
                "description": "Retorna a lista de turmas.",
                "schema": {
                    "type": "array",
                    "items":{
                      "$ref": "#/definitions/Turma"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        },
        "post":{
          "tags": [
            "Turmas"
          ],
          "summary": "Inserir uma nova turma.",
          "description": "Criar uma nova turma.",
          "operationId": "insertTurma",
          "parameters": [
            {
              "name": "turma",
              "in": "formData",
              "description": "Nome da turma.",
              "required": true,
              "type": "string"
            },
            {
              "name": "idprofessor",
              "in": "formData",
              "description": "Código/Username do professor da turma.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "A turma foi criada e é devolvido o seu identificador.",
              "schema": {
                "type": "string",
                "items": {
                    "properties": {
                        "id": {
                            "type": "string"
                        }
                    }
                }
              }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
      "/aplicacoes/turmas/{turma}":{
        "get": {
          "tags": [
            "Turmas"
          ],
          "summary": "Dado o nome da turma, devolve a informação relativa a essa turma.",
          "description": "Devolve a informação relativa a uma turma.",
          "operationId": "getTurmaById",
          "parameters": [
            {
              "name": "turma",
              "in": "path",
              "description": "Nome da turma a retornar.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna a informação relativa a uma turma.",
                "schema": {
                    "$ref": "#/definitions/Turma"   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        },
        "delete":{
            "tags": [
              "Turmas"
            ],
            "summary": "Dado um identifcador da turma, pode apagá-la.",
            "description": "Apagar a informação da turma.",
            "operationId": "deleteTurmaById",
            "parameters": [
              {
                "name": "idTurma",
                "in": "path",
                "description": "Identificador da turma a apagar.",
                "required": true,
                "type": "integer"
              }
            ],
            "responses": {
              "200": {
                  "description": "A turma foi apagada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
      },
      "/aplicacoes/turmas/{turma}/alunos":{
        "get": {
          "tags": [
            "Turmas"
          ],
          "summary": "Dado um identifcador da turma, devolve todos os seus alunos .",
          "description": "Devolve todos os alunos de uma turma.",
          "operationId": "getAlunosByTurma",
          "parameters": [
            {
              "name": "turma",
              "in": "path",
              "description": "Nome da turma.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Lista de alunos da turma.",
                "schema": {
                    "type":"array",
                    "items": {
                      "$ref": "#/definitions/Aluno"
                    }   
                }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      }
    },
    "securityDefinitions":{
      "api_key":{
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "security": [{
       "api_key": []
    }],
    "definitions": {
        "Aluno": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "user": {
                "type": "string"
              },
              "numero": {
                "type": "integer"
              },
              "nome": {
                "type": "string"
              },
              "datanascimento": {
                "type": "string"
              },
              "escola": {
                "type": "string"
              },
              "turma":{
                "type": "string"
              },
              "email":{
                "type":"string"
              },
              "confirmacao":{
                "type":"integer"
              }
            }
        },
...
        "Turma":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "idprofessor": {
              "type": "string"
            },
            "turma":{
              "type":"string"
            }
          }
        },
...
    }   
}